requirements.txt
grpcio-tools==1.18.0
googleapis-common-protos==1.5.6
python-daemon==2.2.0
mysqlclient==1.4.1
SQLAlchemy==1.2.17
python-dotenv==0.10.1
PyYAML==5.1
click==6.7
